Added automatic player information
Automatic player information is polled from a DispatchTimer tick in
MainWindowViewModel using ReadProcessMemory.  Unless I want to inject
code into SC2 to create events (hint: that's a bad idea), this is
probably the best way.  The Tick event itself contains a small FSM.
Offsets is meant to be dynamic; the fields should be changed to readonly
properties that are updated on Offsets.Init.  Offsets and Process should
be properly initialized to avoid invalid states.