Added player name conversion memory
Whenever a player name or race is updated during a game, the underlying
id will be associated with that updated value in an SQLite database and
used for future automatic displaying of name or race.  Building now
requires having the System.Data.SQLite assemblies available from
system.data.sqlite.org.